Remove first load aria-live fiddling
Web specs say aria-live shouldn't be announced when the element with it is first added to the page, so this is guarding against behaviour that aint happening. Might have just been seeing this as a dev tools quirk due to Next auto-updating code changes. Or maybe screen readers are announcing it anyway